updated groups and settings design and structure
