+++ icons/IconGroup.tsx
@@ -1,0 +1,19 @@
+export const IconGroup = ({ size = "2em" }) => {
+  return (
+    <svg
+      width={size}
+      height={size}
+      fill="currentColor"
+      version="1.1"
+      viewBox="0 0 600 550"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g>
+        <path d="m260.78 286.36c-44.469-0.13672-87.176 17.367-118.76 48.668-31.59 31.301-49.477 73.848-49.746 118.31v81.203c0 9.0938 4.8516 17.496 12.727 22.043s17.578 4.5469 25.453 0c7.8789-4.5469 12.73-12.949 12.73-22.043v-81.203c0.19922-30.984 12.703-60.625 34.758-82.391 22.055-21.77 51.852-33.883 82.84-33.68 31.031-0.27344 60.895 11.812 83.004 33.59 22.109 21.773 34.648 51.449 34.852 82.48v81.203c0 9.0938 4.8516 17.496 12.727 22.043s17.578 4.5469 25.453 0c7.8789-4.5469 12.73-12.949 12.73-22.043v-81.203c-0.26953-44.512-18.195-87.098-49.836-118.41-31.645-31.305-74.418-48.777-118.93-48.574z" />
+        <path d="m260.78 241.82c32.082 0.070313 62.875-12.617 85.602-35.262 22.723-22.645 35.516-53.395 35.562-85.477 0.046876-32.082-12.66-62.863-35.324-85.574-22.66-22.711-53.418-35.48-85.5-35.504-32.082-0.023438-62.855 12.707-85.551 35.383-22.691 22.68-35.441 53.445-35.441 85.527 0.066406 32 12.789 62.676 35.395 85.332 22.605 22.652 53.258 35.441 85.258 35.574zm0-190.91v0.003906c18.582-0.070312 36.426 7.2539 49.602 20.352 13.18 13.098 20.609 30.898 20.652 49.477 0.046875 18.582-7.2969 36.414-20.41 49.578-13.113 13.16-30.922 20.57-49.504 20.594-18.578 0.019532-36.406-7.3438-49.551-20.473-13.148-13.133-20.535-30.949-20.535-49.527 0.066406-18.5 7.4297-36.23 20.488-49.336 13.059-13.105 30.758-20.531 49.258-20.664z" />
+        <path d="m438.96 286.36c-9.0938 0-17.5 4.8516-22.047 12.727s-4.5469 17.578 0 25.457c4.5469 7.875 12.953 12.727 22.047 12.727 31.008-0.20313 60.836 11.902 82.93 33.664s34.656 51.398 34.922 82.406v81.203c0 9.0938 4.8516 17.496 12.73 22.043 7.875 4.5469 17.578 4.5469 25.453 0s12.727-12.949 12.727-22.043v-81.203c-0.26953-44.512-18.191-87.098-49.836-118.41-31.641-31.305-74.414-48.777-118.93-48.574z" />
+        <path d="m417.58 54.219c6.9062-2.2344 14.125-3.3516 21.383-3.3086 25.008 0 48.117 13.34 60.621 35 12.504 21.656 12.504 48.34 0 70-12.504 21.656-35.613 35-60.621 35-5.3203-0.10938-10.609-0.79297-15.781-2.0391-8.8242-2.2266-18.164 0.42188-24.504 6.9453-6.3398 6.5273-8.7148 15.938-6.2344 24.691 2.4805 8.7539 9.4414 15.52 18.266 17.746 9.5859 2.3203 19.406 3.5156 29.27 3.5625 35.285-0.22266 68.711-15.848 91.512-42.777 22.801-26.93 32.699-62.477 27.102-97.316-5.6016-34.836-26.141-65.492-56.23-83.918-30.09-18.43-66.73-22.793-100.31-11.949-6.8398 1.6992-12.668 6.1602-16.094 12.316-3.4297 6.1562-4.1523 13.457-2 20.168 2.1523 6.707 6.9922 12.223 13.363 15.234 6.375 3.0078 13.707 3.2422 20.258 0.64453z" />
+      </g>
+    </svg>
+  );
+};
