+++ locales/en.ts
@@ -48,13 +48,14 @@ export const en = {
   groups: {
     group: "Group",
     title: "Groups",
+    manageGroups: "Manage groups",
     backGroupList: "Back to groups list",
     members: "Members",
     createGroupTitle: "Create group",
     groupLabelName: "Group name",
-    createGroupButton: "Create group",
+    createGroupButton: "Create a new group",
     switchGroupButton: "Switch to this group",
-    alreadyCurrentGroup: "already current group",
+    alreadyCurrentGroup: "Already current group",
     notInGroup: "You are not part of any group",
     noMembers: "No members added",
     successGroupCreated: "Group created",
