+++ pages/groups/list.tsx
@@ -1,7 +1,5 @@
 import { useEffect } from "react";
 import type { NextPage } from "next";
-import toast from "react-hot-toast";
-import Link from "next/link";
 import { useAtom } from "jotai";
 
 import LayoutAuth from "../../components/LayoutAuth";
@@ -17,65 +15,58 @@ import {
 import { GroupAtomType } from "../../types/groups";
 import { isLoadingAPIAtom } from "../../state/app";
 import { useAPI } from "../../hooks/useAPI";
+import { GroupListItem } from "../../components/GroupListItem";
+import Button from "../../components/Button";
+import { useRouter } from "next/router";
 
 const GroupList = () => {
   const { t } = useLocale();
 
-  const [groupSession] = useAtom(groupSessionAtom);
   const [groupsList] = useAtom(groupsArrayAtom);
 
-  return !groupsList ? (
-    <p>{t.groups.errorLoadGroups}</p>
-  ) : groupsList.length === 0 ? (
+  if (!groupsList) {
+    return null;
+  }
+
+  return groupsList.length === 0 ? (
     <p>{t.groups.notInGroup}</p>
   ) : (
     groupsList && (
       <ul className={styles.groupList}>
         {groupsList.map((group) => (
-          <li className={styles.groupItem} key={group.id}>
-            <h3>
-              <Link href={`${ROUTES.group}/${group.id}`}>{group.name}</Link>
-              {groupSession?.id === group.id ? ` (${t.common.current})` : ""}
-            </h3>
-            <p className={styles.groupMembers}>
-              <>
-                {`${t.groups.members}: `}
-                {group.members?.length > 0
-                  ? group.members.map((member) => member.user.name).join(", ")
-                  : t.groups.noMembers}
-              </>
-            </p>
-          </li>
+          <GroupListItem key={group.id} group={group} />
         ))}
       </ul>
     )
   );
 };
 
-const Groups: NextPage = () => {
+const GroupsPage: NextPage = () => {
   const { t } = useLocale();
+  const router = useRouter();
 
   const [isLoadingAPI] = useAtom(isLoadingAPIAtom);
   const [, setGroupsList] = useAtom(groupsMapAtom);
   const [groupSession] = useAtom(groupSessionAtom);
 
-  const processSuccess = (data: Array<GroupAtomType>) => {
-    // We display the current group at the top of the list
-    const groupsListMap = data
-      .map((group) => [group.id, group] as [string, GroupAtomType])
-      .sort(([, groupA], [, groupB]) =>
-        groupA.id === groupSession?.id
-          ? -1
-          : groupB.id === groupSession?.id
-          ? 1
-          : 0
-      );
-    setGroupsList(new Map(groupsListMap));
-    return;
-  };
   const { runFetch } = useAPI<Array<GroupAtomType>>({ mode: "manual" });
 
   useEffect(() => {
+    const processSuccess = (data: Array<GroupAtomType>) => {
+      // We display the current group at the top of the list
+      const groupsListMap = data
+        .map((group) => [group.id, group] as [string, GroupAtomType])
+        .sort(([, groupA], [, groupB]) =>
+          groupA.id === groupSession?.id
+            ? -1
+            : groupB.id === groupSession?.id
+            ? 1
+            : 0
+        );
+      setGroupsList(new Map(groupsListMap));
+      return;
+    };
+
     if (groupSession) {
       runFetch({
         endpoint: `${ENDPOINTS.groups}/getAllGroupsBySessionUserId`,
@@ -85,14 +76,26 @@ const Groups: NextPage = () => {
     }
 
     return;
-  }, [t, groupSession]);
+  }, [t, groupSession, setGroupsList, runFetch]);
 
   return (
     <LayoutAuth>
-      <a href={ROUTES.groupsCreate}>Create a new group</a>
-      {isLoadingAPI ? <Spinner /> : <GroupList />}
+      {isLoadingAPI ? (
+        <Spinner />
+      ) : (
+        <>
+          <GroupList />
+          <Button
+            onClick={() => router.push(ROUTES.groupsCreate)}
+            type="blue"
+            isLoading={isLoadingAPI}
+          >
+            {t.groups.createGroupButton}
+          </Button>
+        </>
+      )}
     </LayoutAuth>
   );
 };
 
-export default Groups;
+export default GroupsPage;
