+++ styles/Form.module.scss
@@ -6,7 +6,7 @@
 
   label {
     display: block;
-    margin-bottom: 0.2rem;
+    margin-bottom: 0.4rem;
   }
 
   textarea,
