+++ styles/Layout.module.scss
@@ -42,6 +42,7 @@
 .main {
   flex-grow: 1;
   padding: 1rem 1rem;
+  color: colors.$thirdary;
 }
 
 .footer {
