+++ styles/Settings.module.scss
@@ -1,0 +1,16 @@
+@use "fonts";
+
+.settingsContainer {
+  h2 {
+    font-size: fonts.$bigger;
+    margin-bottom: 1rem;
+  }
+
+  button {
+    margin-bottom: 1rem;
+  }
+
+  p {
+    text-align: center;
+  }
+}
